src/main/webapp/app/entities/tag-type/delete/tag-type-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { TagTypeService } from '../service/tag-type.service';

import { TagTypeDeleteDialogComponent } from './tag-type-delete-dialog.component';

describe('TagType Management Delete Component', () => {
  let comp: TagTypeDeleteDialogComponent;
  let fixture: ComponentFixture<TagTypeDeleteDialogComponent>;
  let service: TagTypeService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, TagTypeDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(TagTypeDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(TagTypeDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(TagTypeService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
